requirements.txt
pandasai>=2.0.0
pandas>=2.1.0
numpy>=1.24.0
streamlit>=1.30.0
python-dotenv>=1.0.1
openpyxl>=3.1.2
PyYAML>=6.0.0
openai>=1.55.0,<2
tabulate>=0.9.0
rapidfuzz>=3.0.0
pyarrow>=14.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
